Fixed EoTDetect.vi
Initially the 'pass' count was reset when EoTDetect was True, the idea being you don't want a false 'EoT' trigger if the next derivative is true. This means the 'EoT' is latched for one call cycle. The problem occurs when the output of this is fed to an AND gate, in which case both would need to latch at the same time.

The concern of a false EoT does not matter since the derivative block is point by point. So the next time called afte EoT = True, the derivative will necessarily be non-zero unless the sensor value has not changed.